Update couriers, kitchen services

+++ delivery-service/src/main/java/ru/liga/dto/CustomerDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CustomerDTO {
     private String address;
-    private int distance;
+    private Double distance;
 }
