+++ delivery-service/src/main/java/ru/liga/dto/RestaurantDTO.java
@@ -2,11 +2,11 @@ package ru.liga.dto;
 
 
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
+
 
 @Data
 public class RestaurantDTO {
     private String address;
-    private int distance;
+    private Double distance;
+
 }
